feat: collect libc types and remove depracated num-derive crate
