+++ crates/libc-types/src/lib.rs
@@ -16,11 +16,13 @@ pub mod elf;
 pub mod epoll;
 pub mod fcntl;
 pub mod futex;
+pub mod ioctl;
 pub mod mman;
 pub mod others;
 pub mod poll;
 pub mod resource;
 pub mod sched;
+pub mod termios;
 pub mod time;
 pub mod times;
 pub mod types;
