+++ filesystem/devfs/Cargo.toml
@@ -11,7 +11,5 @@ devices = { workspace = true }
 log = "0.4"
 sync = { workspace = true }
 bitflags = "2.0.2"
-num-traits = { version = "0.2", default-features = false }
-num-derive = "0.4"
 syscalls = { workspace = true }
 libc-types = { workspace = true }
