+++ filesystem/ramfs/Cargo.toml
@@ -12,3 +12,4 @@ sync = { workspace = true }
 polyhal = { workspace = true }
 runtime = { workspace = true }
 syscalls = { workspace = true }
+libc-types = { workspace = true }
