+++ kernel/Cargo.toml
@@ -35,9 +35,6 @@ futures-lite = { version = "1.13.0", default-features = false, features = [
 hashbrown = "0.14"
 cfg-if = "1.0.0"
 
-num-traits = { version = "0.2", default-features = false }
-num-derive = "0.4"
-
 syscalls = { workspace = true }
 runtime = { workspace = true }
