+++ kernel/src/syscall/time.rs
@@ -14,8 +14,11 @@ use core::{
     task::{Context, Poll},
 };
 use executor::select;
-use fs::TimeSpec;
-use libc_types::{time::ITimerVal, times::TMS, types::TimeVal};
+use libc_types::{
+    time::ITimerVal,
+    times::TMS,
+    types::{TimeSpec, TimeVal},
+};
 use log::{debug, warn};
 use polyhal::time::Time;
 use syscalls::Errno;
